14/Oct/2013 update. Main changes:
Problem with the scope of the objects created that resulted in the Letter objects that were pushed into the alphabet vector being destroyed when exiting the method that did that was fixed.
All the objects are now being correctly passed as reference to the other objects that need to be aware of them (instead of being passed by value, which didn't maintain coherence between what each object was seeing.
the virtual keyboard is now being shown, and depending on the row and column selected each button is shown as active or inactive.